refactor(chain): method-exhaustive PCA classifier + shared failover test harness (#1548 review)
Addresses two otReviewAgent round-4 comments (both behavior/assertion-preserving):

- evm-adapter-conviction: `classifyPcaRead` is now an exhaustive `switch (method)`
  over `PcaRpcMethod` with a `const _exhaustive: never = method` tail, so a newly
  added PCA method must make a conscious routing decision at compile time (TS2322)
  instead of silently falling through to `sticky`. Routing for all six existing
  methods is byte-identical; the now-redundant `PCA_NULLABLE_LOOKUP_METHODS` set is
  removed (the switch encodes nullable routing explicitly). The daemon PCA route
  already allowlists methods before dispatch, so the fail-loud default is
  unreachable in production (strictly safer than the old silent sticky fallback).

- test: the suite split newly cloned the `RpcFailoverClient` harness across both
  sibling files; extract `recorder`, `retryable429`, `NEVER_SIGN`, `makeClient`
  into `rpc-failover-test-helpers.ts` imported by both. Scenario-local doubles
  (readSeq, signer/contract builders, URL sets) stay local. No assertion changed.

Fast lane 600 pass / 1 skip, parity 15, tsc clean, plus an independent two-lens
equivalence review (PCA routing byte-identical; helpers byte-identical, only
import/def lines touched).
