+++ packages/chain/test/endpoint-stickiness.acceptance.unit.test.ts
@@ -16,44 +16,9 @@
  * file (each unit test file stays self-contained).
  */
 import { describe, it, expect, afterEach } from 'vitest';
-import {
-  RpcFailoverClient,
-  type SignPopulatedFn,
-  type StickinessOptions,
-} from '../src/rpc-failover-client.js';
+import { RpcFailoverClient, type SignPopulatedFn } from '../src/rpc-failover-client.js';
 import { _resetRpcFailoverStatsForTest, getRpcFailoverStats } from '../src/rpc-failover-log.js';
-
-function recorder<A extends unknown[], R>(impl: (...args: A) => R) {
-  const calls: A[] = [];
-  const fn = (...args: A): R => { calls.push(args); return impl(...args); };
-  return Object.assign(fn, { calls });
-}
-
-const retryable429 = () => { const e = new Error('429 too many requests'); (e as any).status = 429; return e; };
-
-// A `signPopulated` stub the read/broadcast/receipt families never reach — wired
-// to fail loudly if a non-signing path ever tried to sign.
-const NEVER_SIGN: SignPopulatedFn = async () => {
-  throw new Error('signPopulated must not be reached by this path');
-};
-
-/** Construct the module under test over bare doubles (PLAN §0 D1 thunks).
- *  `stickiness` (Mechanism B) is optional — omitted, the client uses production
- *  defaults (stickiness on unless `DKG_DISABLE_RPC_STICKINESS=1`, `Date.now`
- *  clock, 30s TTL); the stickiness suite injects `{ enabled, now, ttlMs }`. */
-function makeClient(
-  providers: unknown[],
-  rpcUrls: string[],
-  signPopulated: SignPopulatedFn = NEVER_SIGN,
-  stickiness?: StickinessOptions,
-): RpcFailoverClient {
-  return new RpcFailoverClient(
-    () => providers.map((p, i) => ({ provider: p as any, rpcUrl: rpcUrls[i] })),
-    signPopulated,
-    () => 'evm:31337',
-    { stickiness },
-  );
-}
+import { recorder, retryable429, NEVER_SIGN, makeClient } from './rpc-failover-test-helpers.js';
 
 afterEach(() => { _resetRpcFailoverStatsForTest(); });
