+++ packages/chain/test/rpc-failover-client.unit.test.ts
@@ -33,12 +33,7 @@
  * over REAL providers in multi-rpc-{read,write}-failover.test.ts.
  */
 import { describe, it, expect, afterEach, vi } from 'vitest';
-import {
-  RpcFailoverClient,
-  resolveCapMs,
-  type SignPopulatedFn,
-  type StickinessOptions,
-} from '../src/rpc-failover-client.js';
+import { resolveCapMs, type SignPopulatedFn } from '../src/rpc-failover-client.js';
 import {
   RPC_READ_STALL_TIMEOUT_MS,
   RPC_LOG_SCAN_TIMEOUT_MS,
@@ -47,14 +42,8 @@ import {
   RPC_TRANSACTION_POPULATION_ATTEMPT_TIMEOUT_MS,
 } from '../src/evm-adapter-constants.js';
 import { _resetRpcFailoverStatsForTest } from '../src/rpc-failover-log.js';
+import { recorder, retryable429, NEVER_SIGN, makeClient } from './rpc-failover-test-helpers.js';
 
-function recorder<A extends unknown[], R>(impl: (...args: A) => R) {
-  const calls: A[] = [];
-  const fn = (...args: A): R => { calls.push(args); return impl(...args); };
-  return Object.assign(fn, { calls });
-}
-
-const retryable429 = () => { const e = new Error('429 too many requests'); (e as any).status = 429; return e; };
 const callExceptionErr = (msg = 'execution reverted: TooLowAllowance') => {
   const e = new Error(msg); (e as any).code = 'CALL_EXCEPTION'; return e;
 };
@@ -64,30 +53,6 @@ const badDataError = () => {
   return e;
 };
 const knownTxError = () => new Error('already known');
-
-// A `signPopulated` stub the read/broadcast/receipt families never reach — wired
-// to fail loudly if a non-signing path ever tried to sign.
-const NEVER_SIGN: SignPopulatedFn = async () => {
-  throw new Error('signPopulated must not be reached by this path');
-};
-
-/** Construct the module under test over bare doubles (PLAN §0 D1 thunks).
- *  `stickiness` (Mechanism B) is optional — omitted, the client uses production
- *  defaults (stickiness on unless `DKG_DISABLE_RPC_STICKINESS=1`, `Date.now`
- *  clock, 30s TTL); the stickiness suite injects `{ enabled, now, ttlMs }`. */
-function makeClient(
-  providers: unknown[],
-  rpcUrls: string[],
-  signPopulated: SignPopulatedFn = NEVER_SIGN,
-  stickiness?: StickinessOptions,
-): RpcFailoverClient {
-  return new RpcFailoverClient(
-    () => providers.map((p, i) => ({ provider: p as any, rpcUrl: rpcUrls[i] })),
-    signPopulated,
-    () => 'evm:31337',
-    { stickiness },
-  );
-}
 
 const URLS = ['https://primary.example', 'https://backup.example'];
